server/pyproject.toml
[project]
name = "savr-server"
version = "0.1.0"
description = "Savr backend server"
requires-python = ">=3.9"
dependencies = [
    # Core FastAPI dependencies
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "fastapi-cors",
    
    # Database
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    
    # Web scraping and HTTP
    "requests",
    "beautifulsoup4",
    "selenium",
    "webdriver-manager",
    
    # AI and ML
    "openai",
    "pinecone",
    
    # LangChain (essential only)
    "langchain",
    "langchain-openai",
    "langchain-core",
    
    # Image processing
    "opencv-python",
    "Pillow>=10.2.0",
    
    # Date parsing
    "dateparser>=1.2.0",
    "python-dateutil>=2.8.0",
    
    # Environment and configuration
    "python-dotenv",
    
    # Data handling
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    
    # HTTP and async
    "httpx",
    "aiohttp",
    "async-timeout",
    
    # Utilities
    "pydantic",
    "tqdm",
    "typing_extensions",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
